backend/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.8
dnspython==2.3.0
duckduckgo-search==3.8.3
environs==9.5.0
exceptiongroup==1.1.1
fastapi==0.97.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.2
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
isort==5.12.0
langchain==0.0.200
langchainplus-sdk==0.0.10
loguru==0.7.0
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pathspec==0.11.1
pinecone-client==2.2.2
platformdirs==3.5.3
psycopg2-binary==2.9.6
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
SQLAlchemy==2.0.16
starlette==0.27.0
tenacity==8.2.2
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.6.3
urllib3==2.0.3
unstructured==0.7.5
uvicorn==0.22.0
yarl==1.9.2
faiss-cpu==1.7.4
